Moved ensureAdminIfNeeded to new trait SupportsAdmin
